refactor: flatten package structure and remove assets directory
