+++ go-binary/internal/catalog/types.go
@@ -5,7 +5,7 @@ import (
 	"maps"
 	"strings"
 
-	"kubara/internal/assets/service"
+	"kubara/internal/service"
 
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 )
