+++ go-binary/internal/envmap/envMap.go
