+++ go-binary/internal/envmap/envMapMngr.go
