+++ go-binary/internal/envmap/envMapMngr_test.go
