+++ go-binary/internal/service/types.go
